修改启动项目注册配置类，统一抽取通过反射完成注册，添加注册接口和发送验证码接口

+++ flowers-trees-common/src/main/java/com/sj/config/flowerstrees/jwt/JwtConfig.java
@@ -20,7 +20,7 @@ public class JwtConfig {
     /**
      * jwt过期时间 单位秒(s)
      */
-    private Long expire;
+    private Integer expire;
 
     /**
      * jwt前缀
@@ -35,6 +35,6 @@ public class JwtConfig {
     /**
      * jwt过期前多久刷新 单位秒(s)
      */
-    private Long refresh;
+    private Integer refresh;
 
 }
